Fix mocha.cmd binary lookup
This line of code was looking for `mocha.cmd` in Mocha's own `bin` folder.

This is incorrect; there is no `mocha.cmd` file there. The `mocha.cmd` file is in the `node_modules/.bin` folder.

Tested on Windows. Not sure about *nix.

There is probably a better way to get to the `node_modules/.bin` folder than this, but it "works for me (tm)".